12.1铃声选择（尚未完全实现

+++ app/src/main/java/com/example/myapplication/EditScheduleActivity.java
@@ -47,13 +47,11 @@ public class EditScheduleActivity extends AppCompatActivity  implements View.OnC
 
     @Override
     public void onClick(View v){
-        switch (v.getId()){
-            case R.id.deleteSchedule:
-                deleteMySchedule();
-                break;
-            case R.id.editBtn:
-                editSchedule();
-                break;
+        int id = v.getId();
+        if (id == R.id.deleteSchedule) {
+            deleteMySchedule();
+        } else if (id == R.id.editBtn) {
+            editSchedule();
         }
     }
